backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0
email-validator==2.0.0
python-dateutil==2.8.2
pytest==7.4.0
pytest-flask==1.2.0
gunicorn==21.2.0
